fix: os version should fallback to major if no minor exists
